Add binary distribution of lxml
Though lxml is widely available as a Python package, it's not always
easy to install, especially for those with less experience. Therefore,
CtrlShiftGo and I have decided to include a binary distribution of
lxml v3.6.4 with full and appropriate licencing. Given only Python 2,
no other setup is necessary for the script to be run.